+++ R/KG_Color_Schemes.R
@@ -1,16 +1,17 @@
 #' Create complementary color
 #'
-#' \code{show_KG_palette} This function create complementary color.
+#' \code{kg_create_cc} This function create complementary color.
 #' @importFrom grDevices colorRamp rgb
+#'
 #' @param basecols Set basecol color data.
-#' @param alpha	The alpha transparency. Numbers range from 0-1. see argument alpha in \code{\link[grDevices]{rgb}}.
+#' @param alpha	The alpha transparency. Numbers range from 0-1.
+#'      see argument alpha in \code{\link[grDevices]{rgb}}.
 #' @return Return base and complementary color in a data frame
-#' @author Masaki Fujishima \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
-#' @export KG_create_cc
+#' @export
 #' @examples
-#' show_KG_palette(KG_create_cc())
-#' plot_KG_palette(color = KG_create_cc())
-KG_create_cc <- function(basecols = c("#ff48ac", "red"), alpha = 1){
+#' kg_show_consol(kg_create_cc())
+#' kg_plot_color(color = kg_create_cc())
+kg_create_cc <- function(basecols = c("#ff48ac", "red"), alpha = 1){
 
   n <- length(basecols)
   getRGB <- grDevices::colorRamp(colors = basecols, space = "Lab", interpolate = "spline")
@@ -31,17 +32,19 @@ KG_create_cc <- function(basecols = c("#ff48ac", "red"), alpha = 1){
 
 #' Create triad color
 #'
-#' \code{show_KG_palette} This function create triad color.
+#' \code{kg_create_tc} This function create triad color.
 #' @importFrom grDevices colorRamp rgb
+#'
 #' @param basecols Set basecol color data.
-#' @param alpha	The alpha transparency. Numbers range from 0-1. see argument alpha in \code{\link[grDevices]{rgb}}.
+#' @param alpha	The alpha transparency. Numbers range from 0-1.
+#'      see argument alpha in \code{\link[grDevices]{rgb}}.
+#'
 #' @return Return triad color in a data frame
-#' @author Masaki Fujishima \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
-#' @export KG_create_tc
+#' @export
 #' @examples
-#' show_KG_palette(KG_create_tc())
-#' plot_KG_palette(color = KG_create_tc())
-KG_create_tc <- function(basecols = "#4E5C78", alpha = 1){
+#' kg_show_consol(kg_create_tc())
+#' kg_plot_color(color = kg_create_tc())
+kg_create_tc <- function(basecols = "#4E5C78", alpha = 1){
 
   getRGB <- grDevices::colorRamp(color = basecols, space = "Lab", interpolate = "spline")
   create_RGB <- getRGB(seq(0, 1, length.out = 1))/255