+++ NEWS.md
@@ -1,16 +1,19 @@
+# KaradaColor 0.1.5
+Change:Rename some commands to kg_xxx_xxx before allowing CRAN release.
+
 # KaradaColor 0.1.4
-Change:plot_KG_palette() command. Support for color vector data.
+Change:kg_plot_color() command. Support for color vector data.
 Add:convert_colTodf() command. This function convert color data to data.frame class.
 Add:image_add_textAcode() command. This function Add to image() command text and color code.
 
 # KaradaColor 0.1.3
 
-Add:KG_create_cc() command. This function create complementary color.
+Add:kg_create_cc() command. This function create complementary color.
 Change:show_KG_palette() command. Support for data.frame class.
 
 # KaradaColor 0.1.2
 
-Uploaded to CRAN
+Initial release at git.
 
 # KaradaColor 0.1.1